# Use a set to keep track of metric ids for perf gains